Feat : Todo api

See merge request cloud_track/class_04/web_project3/team06/hr-handlers-be!40

+++ src/main/java/com/hr_handlers/vacation/service/VacationService.java
@@ -120,7 +120,7 @@ public class VacationService {
 
     // 잔여 휴가 일수 조회
     public SuccessResponse<VacationBalanceResponse> getBalance(Long employeeId){
-        Long balance = vacationRepository.findEmployeeVacationBalanceById(employeeId);
+        Double balance = vacationRepository.findEmployeeVacationBalanceById(employeeId);
 
         VacationBalanceResponse response = new VacationBalanceResponse(balance);
